% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsiobands.R
\name{plot.iso}
\alias{plot.iso}
\title{Title}
\usage{
\method{plot}{iso}(x, ..., asp = "", add = FALSE)
}
\arguments{
\item{x}{isobands iso object (from \link[isoband:isobands]{isoband::isobands})}

\item{...}{pass arguments to \code{\link[graphics:lines]{graphics::lines()}}}

\item{asp}{aspect ratio}

\item{add}{add to plot or create new one}
}
\value{
nothing, used for side effect (a plot)
}
\description{
Title
}
\examples{
plot(xisobands(volcano))
}
