% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xisoband-package.R
\docType{package}
\name{xisoband-package}
\alias{xisoband}
\alias{xisoband-package}
\title{xisoband: Light Wrapper of 'isoband' to do it the Hypertidy Way}
\description{
Provide use of isoband package using matrix and extent like 'rasterImage', rather than degenerate rectilinear coordinates (like old R image/matlab/netcdf).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
