% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsiobands.R
\name{xisobands}
\alias{xisobands}
\title{Isobands via matix and extent}
\usage{
xisobands(x, extent = NULL, lo = NULL, hi = NULL, nlevs = 12L)
}
\arguments{
\item{x}{matrix of numeric}

\item{extent}{xmin,xmax,ymin,ymax georeference of the x matrix}

\item{lo}{low levels}

\item{hi}{high levels}

\item{nlevs}{number of levels in the case that neither hi nor lo provided}
}
\value{
and isobands, iso object
}
\description{
This is purely a re-orientation of isoband it self, rather than having to construct
a degenerate rectilinear data set, we just provide a  matrix and optionally an
extent. The extent scaling can be done independently if wanted.
}
\details{
Matrix orientation is 'raster', i.e. like \code{\link[=rasterImage]{rasterImage()}} and not like \code{\link[=image]{image()}} which is
isoband is expecting.

Values for lor and/or hi may be provided or left out. If both are left out can use
nlevs to get a quantile set of levels from the data.
}
\examples{
x <- 1:10
y <- 1:5
m <- t(volcano[1:10, 1:5])
#isobands(x, y, m, 101, 106)
i <- 2
## both cases return empty results
isobands(x[i], y, m[, i, drop = FALSE], 101, 106)
isobands(x, y[i], m[i, , drop = FALSE], 101, 106)

m <- volcano
levs <- quantile(m, na.rm = TRUE, probs = seq(0, 1, length.out = 12))
x <- xisobands(m, head(levs, -1), tail(levs, - 1))
plot(x)
}
