% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsiobands.R
\name{plot.iso}
\alias{plot.iso}
\title{Title}
\usage{
\method{plot}{iso}(x, ..., asp = "", add = FALSE)
}
\arguments{
\item{...}{other \link[graphics]{graphical parameters} (see \code{\link[graphics]{par}} and
    section \sQuote{Details} below).}

\item{asp}{the \eqn{y/x} aspect ratio, see \code{\link[graphics]{plot.window}}.}
}
\value{
nothing, used for side effect (a plot)
}
\description{
Title
}
\examples{
plot(xisobands(volcano))
}
