% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsiobands.R
\name{xisobands}
\alias{xisobands}
\title{Isobands via matix and extent}
\usage{
xisobands(x, extent = NULL, lo = NULL, hi = NULL, nlevs = 12L)
}
\arguments{
\item{x}{matrix of numeric}

\item{extent}{xmin,xmax,ymin,ymax georeference of the x matrix}

\item{lo}{low levels}

\item{hi}{high levels}

\item{nlevs}{number of levels in the case that neither hi nor lo provided}
}
\value{
and isobands, iso object
}
\description{
This is purely a re-orientation of isoband it self, rather than having to construct
a degenerate rectilinear data set, we just provide a  matrix and optionally an
extent. The extent scaling can be done independently if wanted.
}
\details{
Matrix orientation is 'raster', i.e. like \code{\link[=rasterImage]{rasterImage()}} and not like \code{\link[=image]{image()}} which is
isoband is expecting.

Values for lor and/or hi may be provided or left out. If both are left out can use
nlevs to get a quantile set of levels from the data.
}
\examples{
xisobands(diag(10))
}
